+++ home_password/blueprints/users/routes.py
@@ -2,8 +2,8 @@ from flask import render_template, url_for, request, redirect, session
 from home_password.models.user import User
 from home_password.models.site import Site
 from flask_login import current_user, login_required
-from flask import Blueprint
-from home_password.blueprints.utils import user_only
+from flask import Blueprint, flash
+from home_password.blueprints.utils import admin_only, user_only
 
 users = Blueprint('users', __name__)
 
@@ -14,10 +14,24 @@ users = Blueprint('users', __name__)
 def home():
   return render_template('users/regular/home.html', sites=current_user.sites)
 
-@users.route("/<int:id>/sites", methods=["GET","POST"])
+
+@users.route('/admin/users/new', methods=["GET","POST"])
 @login_required
-def sites(id):
-  if current_user == User.query.filter_by(id=id).first():
-    return render_template("users/regular/sites.html", all_sites=current_user.sites)
-  else:
-    return redirect(url_for('users.home'))
+@admin_only
+def new_user():
+  if request.method == "POST":
+    print(dict(request.form))
+    if User.query.filter_by(username=request.form["username"]).first() is None:
+      if 'admin' in request.form and 'admin' == request.form["user_type"]:
+        user = User.create_admin(request.form)
+        user.save()
+      else:
+        user = User.create_user(request.form)
+        user.save()
+
+      user.add_sites(request.form.getlist('site'))
+      user.save()
+      flash("User Created", "success")
+      return redirect(url_for("admin.home"))
+    flash("Username in database", "danger")
+  return render_template('users/admin/new_user.html', sites=Site.query.all())
