--- home_password/templates/sites/index.html
@@ -1,16 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-  {% if not not all_sites%}
-    {% for site in all_sites:%}
-      <div class="container">
-        <h1>Site: {{site.name}} </h1>
-        {% if current_user.is_admin %}
-          <h1>Password: {{site.password}}</h1>
-        {% endif %}
-        <a href="/sites/edit/{{site.id}}"><button>Edit</button></a>
-      </div>
-    {% endfor %}
-  {% else %}
-    <h1>There are no sites</h1>
-  {% endif %}
-{% endblock content %}
