--- home_password/templates/users/admin/index.html
@@ -1,14 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-  {% for user in users:%}
-    <div class="container">
-      <h1>{{user.username}}</h1>
-      <a href="/admin/users/edit/{{user.id}}"><button>Edit</button></a>
-      <ul>
-      {% for site in user.sites %}
-        <li>{{site.name}}</li>
-      {% endfor %}
-      </ul>
-    </div>
-  {% endfor %}
-{% endblock content %}
